# Faster `MutatingScope->getNodeKey()`